Minor d_com_static work

+++ include/d/actor/d_a_steam_tag.h
@@ -34,7 +34,7 @@ public:
     /* 0x55C */ dPointWind_c mPointWind;
 
 public:
-    static s32 mEmitterNum;
+    static u8 mEmitterNum;
 };
 
 #endif /* D_A_STEAM_TAG_H */
